requirements.txt
hikari
hikari-lightbulb
uvloop
aiomysql
git+https://github.com/tandemdude/lightbulb-ext-filament